% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcTSP5.R
\name{extractNaturaFormatForMetaModelling}
\alias{extractNaturaFormatForMetaModelling}
\title{Extract plot list for Natura simulation}
\usage{
extractNaturaFormatForMetaModelling(
  QcTSP5Data,
  stratumPlots,
  version = "Natura2014"
)
}
\arguments{
\item{QcTSP5Data}{the database that is retrieved through the restoreQcTSP3Data function}

\item{stratumPlots}{a dataframe with a column "stratum" as a grouping variable as a character field and a column "plots" for plots number as an integer}

\item{version}{a character string identifying the version of the model to be used either "Natura2014" or "RNatura2014"}
}
\value{
a list of two dataframes where the first is a tree list and the second is a list of study trees for model Natura on Capsis Web API
}
\description{
Extract plot list for Natura simulation
}
