% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcTSP5.R
\name{restoreQcTSP5Data}
\alias{restoreQcTSP5Data}
\title{Restore Quebec TSP Data in the Global Environment.}
\usage{
restoreQcTSP5Data()
}
\description{
This function call creates four data.frame objects that contain
the tree measurements.
}
\details{
The resulting list encompasses five data.frame objects are: \cr
\itemize{
\item plots: the list of plots \cr
\item sites: some site information recorded in those plots \cr
\item photoInterpretedStands: some site information recorded through photo-interpretation \cr
\item trees: the tallied trees \cr
\item studyTrees: the study trees (a subsample of trees) \cr
}
}
