% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcTSP5.R
\name{extractArtemisFormatForMetaModelling}
\alias{extractArtemisFormatForMetaModelling}
\title{Extract plot list for Artemis simulation}
\usage{
extractArtemisFormatForMetaModelling(
  QcTSP5Data,
  plots,
  version = "Artemis2009"
)
}
\arguments{
\item{QcTSP5Data}{the database that is retrieved through the restoreQcTSP5Data function}

\item{plots}{a vector of integers standing for the plot id to be considered}
}
\value{
a data.frame object formatted for Capsis Web API
}
\description{
Extract plot list for Artemis simulation
}
