% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcTSP5.R
\name{extractArtemisFormatFromTSP5}
\alias{extractArtemisFormatFromTSP5}
\title{Extract plot list for Artemis simulation}
\usage{
extractArtemisFormatFromTSP5(QcTSP5Data, plots, format = "WebAPI")
}
\arguments{
\item{QcTSP5Data}{the database that is retrieved through the restoreQcTSP5Data function}

\item{plots}{a vector of strings standing for the plot id to be considered}

\item{format}{a character string identifying the intended use, either "WebAPI" or "RPackage"}
}
\value{
a data.frame object formatted for Capsis Web API Capsis or R package simulation
}
\description{
Extract plot list for Artemis simulation
}
\examples{
\dontrun{
 extractArtemisFormatForMetaModelling(QcTSP5Data=Data, 
 plots=c("0300600101", "0300600202",  "0301700302"), 
 format="WebAPI")}

}
